@@ -2,7 +2,7 @@
 
 import React     from 'react';
 import ReactDOM  from 'react-dom';
-import TestUtils from 'react-addons-test-utils';
+import TestUtils from 'react-dom/test-utils';
 import moment    from 'moment';
 import Moment    from '../src/index';
 
